Yutvil/requirements.txt
docx-mailmerge==0.5.0
lxml==4.5.2
Pillow==7.2.0
